Antonio's most Recent Changes
SO i kind of just reference other notes about the 3 copies of the commit
log.  All thought it does not seem right, it is the same as what a lot
of other people have.  Also I did not have much in my notes about the
log record and commit record implementation at the end.  If you have
time maybe you can add to that Chris.  I also did not update the able of
contents yet.  I need to start studying for my final but I can look at
it again tomorrow too.